+++ jest.config.js
@@ -73,6 +73,7 @@ module.exports = {
         '**/packages/{core-common,core-electron-main,core-node,electron-basic,utils}/__tests__/**/?(*.)+(spec|test).[jt]s?(x)',
         // exclude 的要放最后
         '!**/packages/{components,core-browser}/__tests__/**',
+        '!**/packages/extension/__tests__/hosted/**',
       ],
     },
     {
@@ -94,6 +95,7 @@ module.exports = {
         '**/packages/*/__test?(s)__/browser/**/?(*.)+(spec|test).[jt]s?(x)',
         '**/packages/*/__test?(s)__/common/**/?(*.)+(spec|test).[jt]s?(x)',
         '**/tools/*/__tests__/**/?(*.)+(spec|test).[jt]s?(x)',
+        '**/packages/extension/__tests__/hosted/**',
         '**/packages/{components,core-browser,core-common}/__tests__/**/?(*.)+(spec|test).[jt]s?(x)',
       ],
     },
