+++ tools/dev-tool/src/injector-editor.ts
@@ -1,0 +1,81 @@
+import { MonacoService } from '@opensumi/ide-core-browser/lib/monaco';
+import {
+  EditorCollectionService,
+  EditorComponentRegistry,
+  EmptyDocCacheImpl,
+  IEditorDecorationCollectionService,
+  IEditorDocumentModelContentRegistry,
+  IEditorDocumentModelService,
+  IEditorFeatureRegistry,
+  ILanguageService,
+  ResourceService,
+  WorkbenchEditorService,
+} from '@opensumi/ide-editor/lib/browser';
+import { EditorComponentRegistryImpl } from '@opensumi/ide-editor/lib/browser/component';
+import { EditorDocumentModelServiceImpl } from '@opensumi/ide-editor/lib/browser/doc-model/editor-document-model-service';
+import { EditorDocumentModelContentRegistryImpl } from '@opensumi/ide-editor/lib/browser/doc-model/editor-document-registry';
+import { EditorCollectionServiceImpl } from '@opensumi/ide-editor/lib/browser/editor-collection.service';
+import { EditorDecorationCollectionService } from '@opensumi/ide-editor/lib/browser/editor.decoration.service';
+import { EditorFeatureRegistryImpl } from '@opensumi/ide-editor/lib/browser/feature';
+import { LanguageService } from '@opensumi/ide-editor/lib/browser/language/language.service';
+import { ResourceServiceImpl } from '@opensumi/ide-editor/lib/browser/resource.service';
+import { WorkbenchEditorServiceImpl } from '@opensumi/ide-editor/lib/browser/workbench-editor.service';
+import { IDocPersistentCacheProvider } from '@opensumi/ide-editor/lib/common';
+import { IWorkspaceService } from '@opensumi/ide-workspace';
+import { MockWorkspaceService } from '@opensumi/ide-workspace/lib/common/mocks/workspace-service';
+
+import { MockedMonacoService } from '../../../packages/monaco/__mocks__/monaco.service.mock';
+
+import { MockInjector } from './mock-injector';
+export function addEditorProviders(injector: MockInjector) {
+  injector.addProviders(
+    {
+      token: IDocPersistentCacheProvider,
+      useClass: EmptyDocCacheImpl,
+    },
+    {
+      token: IEditorDocumentModelContentRegistry,
+      useClass: EditorDocumentModelContentRegistryImpl,
+    },
+    {
+      token: IEditorDocumentModelService,
+      useClass: EditorDocumentModelServiceImpl,
+    },
+    {
+      token: EditorCollectionService,
+      useClass: EditorCollectionServiceImpl,
+    },
+    {
+      token: WorkbenchEditorService,
+      useClass: WorkbenchEditorServiceImpl,
+    },
+    {
+      token: ResourceService,
+      useClass: ResourceServiceImpl,
+    },
+    {
+      token: EditorComponentRegistry,
+      useClass: EditorComponentRegistryImpl,
+    },
+    {
+      token: IEditorDecorationCollectionService,
+      useClass: EditorDecorationCollectionService,
+    },
+    {
+      token: ILanguageService,
+      useClass: LanguageService,
+    },
+    {
+      token: MonacoService,
+      useClass: MockedMonacoService,
+    },
+    {
+      token: IWorkspaceService,
+      useClass: MockWorkspaceService,
+    },
+    {
+      token: IEditorFeatureRegistry,
+      useClass: EditorFeatureRegistryImpl,
+    },
+  );
+}
